feat(HandlerClient): move command methods in a separate class
